Added print reports functionality on Report display page
Print report function is added to Report page, which is produced
by the Show.js component. To print, react-to-print component is
added (package.json, and yarn.lock updated).  For functional
components, react-to-print requres useRef capability of React.
This is added. The component reference is set to the <div>
encapsulating Formik return. react-to-print requires also trigger
and this is implemented with project Button component. Tested on
Chrome, Firefox and IE11.